refactor(vl): deduplicate MinerU resize logic and simplify coordinate parsing
