+++ oar-ocr-vl/src/doc_parser.rs
@@ -16,7 +16,7 @@
 
 use super::utils::{
     DetectedBox, calculate_overlap_ratio, calculate_projection_overlap_ratio, convert_otsl_to_html,
-    crop_margin, filter_overlap_boxes, text, truncate_repetitive_content,
+    crop_margin, filter_overlap_boxes, image::resize_for_mineru, text, truncate_repetitive_content,
 };
 use image::RgbImage;
 use image::{Rgb, imageops};
@@ -28,7 +28,6 @@ use oar_ocr_core::predictors::LayoutDetectionPredictor;
 use oar_ocr_core::processors::BoundingBox;
 use oar_ocr_core::processors::layout_sorting::sort_layout_enhanced;
 use oar_ocr_core::utils::BBoxCrop;
-use std::borrow::Cow;
 use std::sync::Arc;
 
 /// Recognition task for a layout element.
@@ -681,47 +680,6 @@ impl RecognitionBackend for MinerU {
     fn needs_repetition_truncation(&self) -> bool {
         false // handled inside `recognize()`
     }
-}
-
-/// Resize/pad image for MinerU to meet minimum edge and aspect ratio constraints.
-///
-/// MinerU requires images with minimum edge >= factor (28 by default).
-/// Also pads images with extreme aspect ratios to avoid degenerate inputs.
-fn resize_for_mineru(image: &RgbImage, min_edge: u32, max_aspect_ratio: f32) -> RgbImage {
-    let (mut w, mut h) = image.dimensions();
-    let mut out = Cow::Borrowed(image);
-
-    // Handle extreme aspect ratios by padding
-    let ratio = (w.max(h) as f32) / (w.min(h).max(1) as f32);
-    if ratio > max_aspect_ratio {
-        let (new_w, new_h) = if w > h {
-            (w, (w as f32 / max_aspect_ratio).ceil() as u32)
-        } else {
-            ((h as f32 / max_aspect_ratio).ceil() as u32, h)
-        };
-        let mut canvas = RgbImage::from_pixel(new_w, new_h, Rgb([255, 255, 255]));
-        let x = ((new_w - w) / 2) as i64;
-        let y = ((new_h - h) / 2) as i64;
-        imageops::overlay(&mut canvas, &*out, x, y);
-        out = Cow::Owned(canvas);
-        (w, h) = out.dimensions();
-    }
-
-    // Scale up if minimum edge is too small
-    let min_dim = w.min(h);
-    if min_dim < min_edge {
-        let scale = min_edge as f32 / min_dim as f32;
-        let new_w = (w as f32 * scale).ceil() as u32;
-        let new_h = (h as f32 * scale).ceil() as u32;
-        out = Cow::Owned(imageops::resize(
-            &*out,
-            new_w,
-            new_h,
-            imageops::FilterType::CatmullRom,
-        ));
-    }
-
-    out.into_owned()
 }
 
 fn is_auxiliary_element(element_type: LayoutElementType) -> bool {
